refactor: fold registry-install into SyncableStashProvider.sync() (#125)
Collapse `src/registry-install.ts` (~650 lines) into per-provider
`sync()` implementations on a new `SyncableStashProvider` interface.
This unifies the two parallel fetch pipelines (`installed[]` vs
`stashes[]`) so both `akm add` and `akm update` route through one
provider-factory dispatch.

Key moves:
  - `detectStashRoot()`, `buildInstallCacheDir()`, `applyAkmIncludeConfig()`,
    `downloadArchive()`, `computeFileHash()`, `copyDirectoryContents()`
    → `src/stash-providers/provider-utils.ts`
  - `extractTarGzSecure()`, `validateTarEntries()`, `verifyArchiveIntegrity()`
    → `src/stash-providers/tar-utils.ts` (security-critical, moved intact)
  - npm tarball fetch + extraction → `NpmStashProvider.sync()`
    in new `src/stash-providers/npm.ts`
  - git/github clone pipeline → `GitStashProvider.sync()`
    (github registers as a `GitStashProvider` and rewrites the URL)
  - website snapshot mirror → `WebsiteStashProvider.sync()`
  - filesystem stash → `FilesystemStashProvider.sync()` (no-op)

Orchestration:
  - New `syncFromRef(ref, options)` in `stash-providers/sync-from-ref.ts`
    parses an install ref and dispatches to the correct provider.
  - `akmAdd` now: parse → sync → audit → enforce policy → upsert
    installed entry → write lock entry → akmIndex.
  - `akmUpdate` calls `syncFromRef({ force })` and reruns the audit
    hook, so updates can no longer silently land malicious content.
  - Audit (`auditInstallCandidate`) lives in the orchestrator, not
    inside `sync()`, because the `--trust` decision belongs to the
    caller.

Cleanup:
  - `src/registry-install.ts` deleted; `upsertInstalledRegistryEntry`
    and `removeInstalledRegistryEntry` moved to `stash-add.ts`.
  - `tests/registry-install.test.ts` keeps a local `installRegistryRef`
    helper that calls sync + audit, preserving the existing security
    test suite without resurrecting the deleted module.
