+++ src/stash-providers/index.ts
@@ -7,5 +7,6 @@
  */
 import "./filesystem";
 import "./git";
+import "./npm";
 import "./openviking";
 import "./website";
